+++ src/decoders/index.ts
@@ -1,3 +1,6 @@
 import { decodeSystemInstruction } from "./system";
+import { decodeTokenInstruction } from "./token";
+import { decodeToken2022Instruction } from "./token22";
+import { decodeAssociatedTokenInstruction } from "./associated";
 
-export { decodeSystemInstruction };
+export { decodeSystemInstruction, decodeTokenInstruction, decodeToken2022Instruction, decodeAssociatedTokenInstruction };
